+++ packages/core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Add hovered and selected feature to vuex store. This is an optional functionality that has to be explicitly enabled and works with the `@masterportal/masterportalapi` default marker design. See configuration parameter `extendedMasterportalapiMarkers`.
+- Feature: Add zoomLevel as plugin-agnostic map information to store.
+
 ## 1.1.0
 
 - Feature: Add core state variable for map's center position.